fix(user): add middleware (#19)

+++ server/app/Http/Controllers/UserController.php
@@ -12,11 +12,11 @@ use Illuminate\Routing\Controllers\HasMiddleware;
 
 class UserController extends Controller implements HasMiddleware
 {
-    //Midleware
+
     public static function middleware(): array
     {
         return [
-            // new Middleware('auth:api', only: ['update', 'destroy'])
+            new Middleware('auth:api', except: ['index', 'show'])
         ];
     }
 
@@ -41,8 +41,11 @@ class UserController extends Controller implements HasMiddleware
      */
     public function update(Request $request, string $user)
     {
-
         $user = User::where('slug', $user)->firstOrFail();
+
+        if ($user->id !== $request->user()->id) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
 
         $validated = $request->validate([
             'name' => 'required|string|alpha_dash|min:3|max:50',
@@ -56,7 +59,11 @@ class UserController extends Controller implements HasMiddleware
             ],
             'date_of_birth' => 'nullable|date_format:d-m-Y',
             'bio' => 'nullable|string|max:750',
-            'gender' => [Rule::in(['male', 'female', 'prefer not to say'])],
+            'gender' => [
+                'required',
+                'string',
+                Rule::in(['male', 'female', 'prefer not to say'])
+            ],
         ]);
 
         $user->fill($validated);
@@ -68,9 +75,14 @@ class UserController extends Controller implements HasMiddleware
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(string $user)
+    public function destroy(Request $request, string $user)
     {
         $user = User::where('slug', $user)->firstOrFail();
+
+        if ($user->id !== $request->user()->id) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
+
         $user->delete();
 
         return response()->json(['message' => 'User has been deleted successfully'], 204);
